+++ src/kernel.cpp
@@ -1,0 +1,97 @@
+#include <stdint.h>
+#include <stddef.h>
+#include "stivale2.h"
+#include "kprintf.h"
+
+// We need to tell the stivale bootloader where we want our stack to be.
+// We are going to allocate our stack as an uninitialised array in .bss.
+static uint8_t stack[4096];
+
+// stivale2 uses a linked list of tags for both communicating TO the
+// bootloader, or receiving info FROM it. More information about these tags
+// is found in the stivale2 specification.
+
+
+static struct stivale2_header_tag_terminal terminal_hdr_tag = {
+        .tag ={
+            .identifier = STIVALE2_HEADER_TAG_TERMINAL_ID,
+            .next = 0
+        },
+
+        .flags = 0
+};
+
+
+// As an example header tag, we're gonna define a framebuffer header tag.
+// This tag tells the bootloader that we want a graphical framebuffer instead
+// of a CGA-compatible text mode. Omitting this tag will make the bootloader
+// default to text mode.
+static struct stivale2_header_tag_framebuffer framebuffer_hdr_tag = {
+    // All tags need to begin with an identifier and a pointer to the next tag.
+    .tag = {
+        // Identification constant defined in stivale2.h and the specification.
+        .identifier = STIVALE2_HEADER_TAG_FRAMEBUFFER_ID,
+        // If next is 0, then this marks the end of the linked list of tags.
+        .next = (uint64_t)&terminal_hdr_tag
+    },
+    // We set all the framebuffer specifics to 0 as we want the bootloader
+    // to pick the best it can.
+    .framebuffer_width  = 0,
+    .framebuffer_height = 0,
+    .framebuffer_bpp    = 0
+};
+
+
+__attribute__((section(".stivale2hdr"), used))
+static struct stivale2_header stivale_hdr = {
+
+    .entry_point = 0,
+
+    .stack = (uintptr_t)stack + sizeof(stack),
+
+    .flags = 0,
+
+    .tags = (uintptr_t)&framebuffer_hdr_tag
+};
+
+void *stivale2_get_tag(struct stivale2_struct *stivale2_struct, uint64_t id) {
+    struct stivale2_tag *current_tag = (struct stivale2_tag *)stivale2_struct->tags;
+    for (;;) {
+        if (current_tag == NULL) {
+            return NULL;
+        }
+        if (current_tag->identifier == id) {
+            return current_tag;
+        }
+        current_tag = (stivale2_tag *)current_tag->next;
+    }
+}
+
+void _start(struct stivale2_struct *stivale2_struct) {
+    struct stivale2_struct_tag_framebuffer *fb_str_tag;
+    fb_str_tag = (stivale2_struct_tag_framebuffer*)stivale2_get_tag(stivale2_struct, STIVALE2_STRUCT_TAG_FRAMEBUFFER_ID);
+
+    struct stivale2_struct_tag_terminal *term_str_tag;
+    term_str_tag = (stivale2_struct_tag_terminal*)stivale2_get_tag(stivale2_struct, STIVALE2_STRUCT_TAG_TERMINAL_ID);
+
+    if (fb_str_tag == NULL) {
+
+        for (;;) {
+            asm ("hlt");
+        }
+    }
+
+    if(term_str_tag == NULL) {
+        for(;;) {
+            asm ("hlt");
+        }
+    }
+
+    uint8_t *fb_addr = (uint8_t *)fb_str_tag->framebuffer_addr;
+    stivale2_term_write = (void(*)(const char*, size_t))term_str_tag->term_write;
+    kprintf("Welcome to cosmOS! Work in progress");
+
+    for (;;) {
+        asm ("hlt");
+    }
+}
